Fix build issues and tests
