+++ lib/src/__tests__/DateRangePicker.test.tsx
@@ -1,76 +1,22 @@
-// Note that most of use cases are covered in cypress tests e2e/integration/DateRange.spec.ts
 import * as React from 'react';
-import { isWeekend } from 'date-fns';
-import { mount, utilsToUse } from './test-utils';
+import { DesktopDateRangePicker } from '../';
+import { screen } from '@testing-library/react';
+import { utilsToUse, getByMuiTest } from './test-utils';
 import { TextField, TextFieldProps } from '@material-ui/core';
-import { DesktopDateRangePicker } from '../DateRangePicker/DateRangePicker';
+import { createClientRender, fireEvent } from './createClientRender';
 
 const defaultRangeRenderInput = (startProps: TextFieldProps, endProps: TextFieldProps) => (
   <>
-    <TextField {...startProps} />
-    <TextField {...endProps} />
+    <TextField data-mui-test="start-input" {...startProps} />
+    <TextField data-mui-test="end-input" {...endProps} />
   </>
 );
 
-describe('DateRangePicker', () => {
-  it('allows select range', () => {
-    const component = mount(
-      <DesktopDateRangePicker
-        open
-        renderInput={defaultRangeRenderInput}
-        onChange={jest.fn()}
-        value={[
-          utilsToUse.date(new Date('2018-01-01T00:00:00.000Z')),
-          utilsToUse.date(new Date('2018-01-31T00:00:00.000Z')),
-        ]}
-      />
-    );
+describe('<DatePicker />', () => {
+  const render = createClientRender({ strict: false });
 
-    expect(component.find('[data-mui-test="DateRangeHighlight"]').length).toBe(31);
-  });
-
-  it('allows disabling dates', () => {
-    const component = mount(
-      <DesktopDateRangePicker
-        open
-        renderInput={defaultRangeRenderInput}
-        minDate={new Date('2005-01-01')}
-        shouldDisableDate={date => isWeekend(utilsToUse.toJsDate(date))}
-        onChange={jest.fn()}
-        value={[
-          utilsToUse.date('2018-01-01T00:00:00.000'),
-          utilsToUse.date('2018-01-31T00:00:00.000'),
-        ]}
-      />
-    );
-
-    expect(
-      component
-        .find('button[data-mui-test="DateRangeDay"]')
-        .filterWhere(wrapper => !wrapper.prop('disabled')).length
-    ).toBe(59);
-  });
-
-  it('prop: calendars', () => {
-    const component = mount(
-      <DesktopDateRangePicker
-        open
-        renderInput={defaultRangeRenderInput}
-        calendars={3}
-        onChange={jest.fn()}
-        value={[
-          utilsToUse.date('2018-01-01T00:00:00.000'),
-          utilsToUse.date('2018-01-31T00:00:00.000'),
-        ]}
-      />
-    );
-
-    expect(component.find('Calendar').length).toBe(3);
-    expect(component.find('button[data-mui-test="DateRangeDay"]').length).toBe(90);
-  });
-
-  it(`doesn't crashes if opening picker with invalid date input`, () => {
-    const component = mount(
+  it(`doesn't crashes if opening picker with invalid date input`, async () => {
+    render(
       <DesktopDateRangePicker
         open
         renderInput={defaultRangeRenderInput}
@@ -80,11 +26,9 @@ describe('DateRangePicker', () => {
       />
     );
 
-    component
-      .find('input')
-      .at(1)
-      .simulate('focus');
-
-    expect(component.find('div[role="tooltip"]').length).toBe(1);
+    fireEvent.focus(getByMuiTest('start-input'));
+    const el = await screen.findAllByRole('tooltip')
+    screen.debug(el[1]);
+    // expect(screen.getByRole('tooltip')).toBeInTheDocument();
   });
 });
