--- lib/src/_shared/CanAutoFocusContext.tsx
@@ -1,5 +1,0 @@
-import * as React from 'react';
-
-export const CanAutoFocusContext = React.createContext(true);
-
-export const useCanAutoFocus = () => React.useContext(CanAutoFocusContext);
